apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
supabase==2.10.0
google-cloud-aiplatform==1.78.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.24.0
